pyproject.toml
[project]
name = "spec-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.21.0",
    "google-genai>=1.56.0",
    "jieba>=0.42.1",
    "litellm>=1.80.12",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "rank-bm25>=0.2.2",
]
